# Experimental: Compute and cache transitive digests as files in build_resolvers