--- a/_test/test/goldens/generated_build_script.dart
+++ b/_test/test/goldens/generated_build_script.dart
@@ -7,9 +7,10 @@ import 'package:build_test/builder.dart' as _i4;
 import 'package:build_config/build_config.dart' as _i5;
 import 'package:build_modules/builders.dart' as _i6;
 import 'package:build/build.dart' as _i7;
-import 'dart:isolate' as _i8;
-import 'package:build_runner/build_runner.dart' as _i9;
-import 'dart:io' as _i10;
+import 'package:build_resolvers/builder.dart' as _i8;
+import 'dart:isolate' as _i9;
+import 'package:build_runner/build_runner.dart' as _i10;
+import 'dart:io' as _i11;
 
 final _builders = <_i1.BuilderApplication>[
   _i1.apply(
@@ -129,6 +130,13 @@ final _builders = <_i1.BuilderApplication>[
         const _i7.BuilderOptions(<String, dynamic>{r'compiler': r'dart2js'}),
     appliesBuilders: const [r'build_web_compilers:dart2js_archive_extractor'],
   ),
+  _i1.apply(
+    r'build_resolvers:transitive_digests',
+    [_i8.transitiveDigestsBuilder],
+    _i1.toAllPackages(),
+    isOptional: true,
+    hideOutput: true,
+  ),
   _i1.applyPostProcess(
     r'build_modules:module_cleanup',
     _i6.moduleCleanup,
@@ -152,12 +160,12 @@ final _builders = <_i1.BuilderApplication>[
 ];
 void main(
   List<String> args, [
-  _i8.SendPort? sendPort,
+  _i9.SendPort? sendPort,
 ]) async {
-  var result = await _i9.run(
+  var result = await _i10.run(
     args,
     _builders,
   );
   sendPort?.send(result);
-  _i10.exitCode = result;
+  _i11.exitCode = result;
 }