--- a/build_resolvers/pubspec.yaml
+++ b/build_resolvers/pubspec.yaml
@@ -11,6 +11,7 @@ dependencies:
   async: ^2.5.0
   build: ^2.0.0
   collection: ^1.17.0
+  convert: ^3.1.1
   crypto: ^3.0.0
   graphs: '>=1.0.0 <3.0.0'
   logging: ^1.0.0