… build_resolvers/CHANGELOG.md
… build_resolvers/build.yaml
… build_resolvers/lib/src/shared_resource_pool.dart
… docs/faq.md