@jakemac53 2023-08-09
I did quickly run this on the `build_modules` package which uses json_serializable and it was slower than the current build, at least initially. But that isn't a very representative example.

I would love to get some initial community feedback on it, you should be able to try it by just adding the following to your pubspec:

```
dependency_overrides:
  build_resolvers:
    git:
      url: https://github.com/dart-lang/build.git
      ref: transitive_deps
```

If you do try this out, please tell me how it works for you as well as your project structure. Do you have a single large app package or multiple sub-packages? Roughly how many files do you have per package?

cc @simolus3 @rrousselGit not sure if you have some more complex examples worth running this on