@natebosch 2023-08-09
> it was slower than the current build

Where would the extra time be going? Did you compare to a full build after a clean, or to an incremental build?