@jakemac53 2023-08-09
> Where would the extra time be going? Did you compare to a full build after a clean, or to an incremental build?

It was in spent creating these files - I only did clean builds because I was editing build_modules where incremental builds are impossible.