@jakemac53 2023-08-09
I tried the example project in json_serializable and build times are just slightly longer on the intial build and basically identical for other builds - but that is also not a good example because there are very few deps.

I think we need some data from real world projects here, or we could make some example setups to see how things perform in the worst and best case scenarios.