@simolus3 2023-08-09
Would it still be sound if `PerActionResolver` also called `reportUnusedAssets()` on all assets read for the analyzer when the transitive digest is available? Perhaps that would be a more aggressive optimization to reduce dependecies in the asset graph.

I'm happy to run this on the examples in the drift repository, but I suppose they aren't a good indicator either. Applying the optimization to [this larger project](https://github.com/syphon-org/syphon/) (my change to test incremental rebuilds was to edit a comment in `lib/domain/events/messages/schema.dart`) was inconclusive too.