@ikbendewilliam 2023-08-10
In the project I'm currently working on, we have 968 dart files in lib and 207 in test. It's one large app package. We have 6 generator packages (I think 8 generators). 
The command I used: `fvm flutter packages pub run build_runner build --delete-conflicting-outputs`
(I do need to optimise our generate-for)

Initial build without this experimental PR: `Succeeded after 6m 17s with 562 outputs (10219 actions)`
Initial build with it: `Succeeded after 7m 19s with 3752 outputs (13409 actions)`

Adding a single argument in a viewmodel results in the following builds:
Without: `Succeeded after 2m 13s with 90 outputs (625 actions)`
With: `Succeeded after 3m 7s with 90 outputs (625 actions)`

Removing the same argument to see it wasn't a fluke:
Without: `Succeeded after 2m 5s with 90 outputs (625 actions)`
With: `Succeeded after 2m 55s with 90 outputs (625 actions)`

Hope this helps, if you want I have more projects to run on and can do different changes, but this one is the only one where it takes longer.