@jakemac53 2023-08-10
> Would it still be sound if `PerActionResolver` also called `reportUnusedAssets()` on all assets read for the analyzer when the transitive digest is available? Perhaps that would be a more aggressive optimization to reduce dependecies in the asset graph.

If you mean the files read because they haven't yet been loaded into the analyzer, then yes it would be sound to call `reportUnusedAssets` but it wouldn't make much of a difference other than in asset graph size. It shouldn't change the actual invalidation behavior to have the extra dependencies since the transitive digest file would capture the same information.