@jakemac53 2023-08-10
> In the project I'm currently working on, we have 968 dart files in lib and 207 in test. It's one large app package. We have 6 generator packages (I think 8 generators).

Thanks for the information and trying it out. It isn't too surprising then that it isn't helping based on that structure. The optimization only really applies to your package dependencies by default, you would have to split the app package up into smaller packages or try adding multiple targets in your build.yaml, which can be a hard thing to get right (and keep right) depending on your structure of your app. If you are interested in getting help doing that I could help you, but if not I totally understand.

I think the extra async work we are doing to check if these files are readable outweighing any advantages you are getting currently.

I am going to see if I can write some tests to validate that the optimization is in fact actually working and I didn't just screw up the implementation somehow, because I am surprised overall by the behavior we are all experiencing.