@jakemac53 2023-08-10
Ok, I have identified some problem areas in the implementation as well as some optimization opportunities. I don't have a _ton_ of time to spend on this but given the potential impact for people I think its worth spending a bit more effort here. I will follow up in a bit and probably ask people to try again :).