@jakemac53 2023-08-10
ok @ikbendewilliam can you try again with the latest? I think you will need to run another `[dart|flutter] pub get` to get the latest version of the branch.

In my testing for devtools, incremental builds are about twice as fast now, although I did add two custom targets there to realize that benefit, but even without that it was about 30% faster than before and initial builds were also about 10% faster.