@jakemac53 2023-08-11
Ok, I built a fake app, just to try and get a sense of how this can scale. The structure isn't likely very representative, but there are a large number of files (>10k), enough that I knew the old approach would not scale well. Any app _could_ be structured in roughly this way and actually this structure is somewhat intentionally bad (for the original approach at least). We know we have apps that have roughly this number of files (or even greater) in the wild. Although they likely aren't doing analyzer based codegen on every single library like this.

Specifically the structure is 100 packages with 101 files each. Each package has 1 entrypoint in lib which imports all the other 100 files. Those files are under `lib/src` with numbers like `0.dart` etc. Those all import all previous ones, so `2.dart` imports `1.dart` and `0.dart`.

Each file (the entrypoint as well as lib/src files)  also imports the entrypoint of each previous app, so files in `pkg5` import `package:pkg4/pkg4.dart` through `package:pkg0/pkg0.dart`.

I then have a root package which is hand written and defines a single builder, which applies to all packages, and all dart files. It resolves each file, and outputs a simple text file with the list of top level declarations in each file (there are actually none, so it doesn't really do anything other than resolve each library).

That root package also depends on pkg99, so that codegen will run on all those packages.

With this change, I got the following results:

1m 49s fresh (11601 outputs, 11669 actions)
1m 48s editing pkg50/pkg50.dart (4999 outputs, 4999 actions) - **Observation**: why is this almost as long as a fresh build?
56.5s editing pkg75/pkg75.dart (2449 outputs, 2449 actions)
22.6s editing pkg90/pkg90.dart (919 outputs,919  actions)
10.6s editing pkg95/pkg95.dart (409 outputs, 409 actions)
627ms editing pkg99/pkg99.dart (1 outputs, 1  actions)

Without this change, I get:

26m 14s fresh (11361 outputs, 11429 actions)
 editing pkg50/pkg50.dart ( outputs,  actions)
 editing pkg75/pkg75.dart ( outputs,  actions)
 editing pkg90/pkg90.dart ( outputs,  actions)
 editing pkg95/pkg95.dart ( outputs,  actions)
 editing pkg99/pkg99.dart ( outputs,  actions)
 
 (will fill in above results as they complete, its a long wait LOL)