@ikbendewilliam 2023-08-11
I've done a similar test as last time, here are the results:

The initial build FAILED with following error:
```dart
[SEVERE] drift_dev on test/database/people/people_dao_storage_test.dart:

Invalid argument(s): Missing library: package:flutter_test/flutter_test.dart
```
(with a long list of packages)
Without the PR it does work!

and another one I saw:
```dart
[SEVERE] drift_dev on lib/some/viewmodel/some_viewmodel.dart:

LateInitializationError: Field 'library' has not been initialized.
```

Initial build: `4m 7s, 10607 actions` (without PR `7m 40s, 562 outputs & 10219 actions`)
Adding an argument in a VM: `1m 18s-1m 30s, 1257 actions` (without PR `2m 20s, 90 outputs & 625 actions`)

So definetly faster this time! But it seems there is an issue that breaks drift_dev

If you want I can reach out for a project with multiple packages.