@greglittlefield-wf 2023-08-14
> cc @robrbecker I am not sure if you all still use any codegen that uses the Resolver or just rely on parsing nowadays. If you do though, you should give PR a try and see how it does!

We do still have a few builders that use Resolver, such as built_value and json_serializable, so we gave these changes a spin. Thanks for putting this experiment together and letting us know about it! 

I wasn't able to pull in this branch specifically, but I attempted to backpatch the changes to our version build_resolvers 1.5.4 ([backpatch branch](https://github.com/greglittlefield-wf/build/compare/build_resolvers-v1.5.4...greglittlefield-wf:build:transitive-digests--build_resolvers-v1.5.4?expand=1)) and that seemed to build alright, and did end up impacting performance, so that's what we tested with.

**tl;dr** Generally, we didn't see huge improvements except for when I removed the existing `generate_for` configurations that restricted the files that resolved-AST builders run on, except for one case where we saw a modest speedup.

<details>
<summary><em>Expand for more details</em></summary>

#### Large application

I first tried testing this change on our largest application, which has little code of its own and many dependencies.

Locally, on an M1 mac, for a fresh `pub run build_runner build`:

- Before: 5m 3s
- With this build_resolvers patch: 4m 43s

So, a speedup of 20s, taking as long as 93.4% of the original build time.

In our CI environment, however, I saw a slight slowdown:
- Without patch: 9m 51s, 9m 45s
- With patch: 10m 0s, 10m 8s

A slowdown of 16.0s, or 103% of the original time.

#### Medium-sized package

I tested in one of the application's dependencies that had more of its own code, and didn't see a significant difference in build times. 

However, I remembered that this package, along with most of our packages, have `generate_for` filters for its resolved-AST builders so they only resolve the files they needed to, as a performance optimization.

So, I tried taking those out and seeing the difference the patch made.

Before taking those filters out, locally I was seeing:
-  2m 43s for a fresh build
-  8.9s for a rebuild while running `pub run build_runner watch`, after changing a specific file that utilized a resolved-AST builder and was imported in a lot of other places

Without `generate_for` filters, locally I was seeing:
- Without build_resolvers patch: 
    - 6m 34s initial build
    - 2m 43s `watch` rebuild after changing the same file as above
- With build_resolvers patch: 
    - 4m 27s Initial build
    - 52.1s `pbr watch` rebuild after changing the same file as above

So, the patch sped up:
- initial builds up by 2m 7s, taking as long as 67.8% the time without the patch
- rebuilds by 1m 50.9s , taking as long as 32.0% the time without the patch

However, things were still much slower than when we had those `generate_for`s in place.

#### Other packages

We tried the same in a couple other packages, but generally saw about the same performance with/without the patch.

However, there was a case of a larger package where we saw the initial build go from 9m 5s to 7m 12s locally.

</details>